+++ core/com.b2international.snowowl.datastore.server/src/com/b2international/snowowl/datastore/server/cdo/AddedInSourceAndTargetConflict.java
@@ -28,17 +28,11 @@ public class AddedInSourceAndTargetConflict extends Conflict {
 	private final CDOID sourceId;
 	private final CDOID targetId;
 	private final String message;
-	private final boolean addedInSource;
-
-	public AddedInSourceAndTargetConflict(final CDOID sourceId, final CDOID targetId, final String message) {
-		this(sourceId, targetId, message, true);
-	}
 	
-	public AddedInSourceAndTargetConflict(final CDOID sourceId, final CDOID targetId, final String message, final boolean addedInSource) {
+	public AddedInSourceAndTargetConflict(final CDOID sourceId, final CDOID targetId, final String message) {
 		this.sourceId = sourceId;
 		this.targetId = targetId;
 		this.message = message;
-		this.addedInSource = addedInSource;
 	}
 
 	@Override
@@ -56,10 +50,6 @@ public class AddedInSourceAndTargetConflict extends Conflict {
 
 	public CDOID getTargetId() {
 		return targetId;
-	}
-	
-	public boolean isAddedInSource() {
-		return addedInSource;
 	}
 	
 	@Override
