+++ core/com.b2international.snowowl.datastore.server/src/com/b2international/snowowl/datastore/server/internal/branch/CDOBranchMerger.java
@@ -16,7 +16,6 @@
 package com.b2international.snowowl.datastore.server.internal.branch;
 
 import java.util.Collection;
-import java.util.Map;
 
 import org.eclipse.emf.cdo.common.id.CDOID;
 import org.eclipse.emf.cdo.common.revision.CDORevision;
@@ -28,10 +27,7 @@ import org.eclipse.emf.ecore.EStructuralFeature;
 import org.eclipse.emf.spi.cdo.DefaultCDOMerger;
 
 import com.b2international.snowowl.core.merge.MergeConflict;
-import com.b2international.snowowl.datastore.server.cdo.ConflictMapper;
 import com.b2international.snowowl.datastore.server.cdo.ICDOConflictProcessor;
-import com.google.common.base.Function;
-import com.google.common.collect.Maps;
 
 /**
  * An extension of CDO's {@link ManyValued many-valued} merger implementation, that delegates to a terminology-specific
@@ -40,45 +36,43 @@ import com.google.common.collect.Maps;
 public class CDOBranchMerger extends DefaultCDOMerger.PerFeature.ManyValued {
 
 	private final ICDOConflictProcessor delegate;
-	private final boolean isRebase;
 
 	/**
 	 * Creates a new instance using the specified repository identifier.
 	 * 
 	 * @param delegate the CDO conflict processor handling terminology-specific merge decisions 
 	 */
-	public CDOBranchMerger(final ICDOConflictProcessor delegate, final boolean isRebase) {
+	public CDOBranchMerger(final ICDOConflictProcessor delegate) {
 		this.delegate = delegate;
-		this.isRebase = isRebase;
 	}
 
 	@Override
-	protected Object addedInSource(final CDORevision revision) {
+	protected final Object addedInSource(final CDORevision revision) {
 		return delegate.addedInSource(revision, getTargetMap());
 	}
 
 	@Override
-	protected Object addedInTarget(final CDORevision revision) {
-		return delegate.addedInTarget(revision, getSourceMap());
+	protected final Object addedInTarget(final CDORevision revision) {
+		throw new IllegalStateException("Dead code");
 	}
 
 	@Override
-	protected Object detachedInSource(CDOID id) {
+	protected final Object detachedInSource(CDOID id) {
 		return delegate.detachedInSource(id);
 	}
 	
 	@Override
-	protected Object detachedInTarget(CDOID id) {
-		return delegate.detachedInTarget(id);
+	protected final Object detachedInTarget(CDOID id) {
+		throw new IllegalStateException("Dead code");
 	}
 	
 	@Override
-	protected Object changedInTargetAndDetachedInSource(final CDORevisionDelta targetDelta) {
+	protected final Object changedInTargetAndDetachedInSource(final CDORevisionDelta targetDelta) {
 		return delegate.changedInTargetAndDetachedInSource(targetDelta);
 	}
 	
 	@Override
-	protected CDOFeatureDelta changedInSourceAndTargetSingleValued(final EStructuralFeature feature, 
+	protected final CDOFeatureDelta changedInSourceAndTargetSingleValued(final EStructuralFeature feature, 
 			final CDOFeatureDelta targetFeatureDelta, 
 			final CDOFeatureDelta sourceFeatureDelta) {
 
@@ -87,30 +81,14 @@ public class CDOBranchMerger extends DefaultCDOMerger.PerFeature.ManyValued {
 	
 	@Override
 	protected void preProcess() {
-		delegate.preProcess(getSourceMap(), getTargetMap(), isRebase);
-	}
-
-	@Override
-	public Map<CDOID, Conflict> getConflicts() {
-		// Due to the nature of rebase we need to transform certain conflicts to reflect the source and target branches properly
-		if (isRebase) {
-			return Maps.transformValues(super.getConflicts(), new Function<Conflict, Conflict>() {
-				@Override public Conflict apply(Conflict input) {
-					return ConflictMapper.invert(input);
-				}
-			});
-		}
-		return super.getConflicts();
+		delegate.preProcess(getSourceMap(), getTargetMap());
 	}
 	
 	public void postProcess(final CDOTransaction transaction) {
 		delegate.postProcess(transaction);
 	}
 	
-	public Collection<MergeConflict> handleCDOConflicts(final CDOView sourceView, final CDOView targetView) {
-		if (isRebase) {
-			return delegate.handleCDOConflicts(targetView, sourceView, getConflicts());
-		}
-		return delegate.handleCDOConflicts(sourceView, targetView, getConflicts());
+	public Collection<MergeConflict> handleCDOConflicts(final CDOView sourceView, final CDOView targetView, boolean invertConflicts) {
+		return delegate.handleCDOConflicts(sourceView, targetView, getConflicts(), invertConflicts);
 	}
 }
