+++ snomed/com.b2international.snowowl.snomed.datastore.server/src/com/b2international/snowowl/datastore/server/snomed/SnomedCDOConflictProcessor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2016 B2i Healthcare Pte Ltd, http://b2i.sg
+ * Copyright 2011-2017 B2i Healthcare Pte Ltd, http://b2i.sg
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,6 +16,9 @@
 package com.b2international.snowowl.datastore.server.snomed;
 
 import java.util.Collection;
+import java.util.Collections;
+import java.util.Date;
+import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
@@ -25,23 +28,35 @@ import org.eclipse.emf.cdo.common.id.CDOID;
 import org.eclipse.emf.cdo.common.revision.CDORevision;
 import org.eclipse.emf.cdo.common.revision.delta.CDOFeatureDelta;
 import org.eclipse.emf.cdo.common.revision.delta.CDOFeatureDelta.Type;
+import org.eclipse.emf.cdo.common.revision.delta.CDORevisionDelta;
+import org.eclipse.emf.cdo.common.revision.delta.CDOSetFeatureDelta;
 import org.eclipse.emf.cdo.spi.common.revision.InternalCDORevision;
+import org.eclipse.emf.cdo.util.ObjectNotFoundException;
 import org.eclipse.emf.cdo.view.CDOView;
 import org.eclipse.emf.ecore.EAttribute;
 import org.eclipse.emf.ecore.EClass;
 import org.eclipse.emf.ecore.EStructuralFeature;
+import org.eclipse.emf.spi.cdo.DefaultCDOMerger.ChangedInSourceAndDetachedInTargetConflict;
+import org.eclipse.emf.spi.cdo.DefaultCDOMerger.ChangedInSourceAndTargetConflict;
+import org.eclipse.emf.spi.cdo.DefaultCDOMerger.ChangedInTargetAndDetachedInSourceConflict;
 import org.eclipse.emf.spi.cdo.DefaultCDOMerger.Conflict;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.b2international.commons.Pair;
+import com.b2international.snowowl.core.date.DateFormats;
+import com.b2international.snowowl.core.date.Dates;
+import com.b2international.snowowl.core.merge.ConflictingAttribute;
+import com.b2international.snowowl.core.merge.ConflictingAttributeImpl;
 import com.b2international.snowowl.core.merge.MergeConflict;
+import com.b2international.snowowl.core.merge.MergeConflict.ConflictType;
+import com.b2international.snowowl.core.merge.MergeConflictImpl;
 import com.b2international.snowowl.datastore.server.cdo.AbstractCDOConflictProcessor;
 import com.b2international.snowowl.datastore.server.cdo.AddedInSourceAndDetachedInTargetConflict;
 import com.b2international.snowowl.datastore.server.cdo.AddedInSourceAndTargetConflict;
 import com.b2international.snowowl.datastore.server.cdo.AddedInTargetAndDetachedInSourceConflict;
 import com.b2international.snowowl.datastore.server.cdo.ICDOConflictProcessor;
-import com.b2international.snowowl.datastore.server.snomed.merge.SnomedMergeConflictMapper;
+import com.b2international.snowowl.snomed.Component;
 import com.b2international.snowowl.snomed.Relationship;
 import com.b2international.snowowl.snomed.SnomedPackage;
 import com.b2international.snowowl.snomed.datastore.SnomedDatastoreActivator;
@@ -49,6 +64,7 @@ import com.b2international.snowowl.snomed.snomedrefset.SnomedRefSetMember;
 import com.b2international.snowowl.snomed.snomedrefset.SnomedRefSetPackage;
 import com.google.common.base.Function;
 import com.google.common.base.Predicate;
+import com.google.common.base.Strings;
 import com.google.common.collect.FluentIterable;
 import com.google.common.collect.HashMultimap;
 import com.google.common.collect.ImmutableMap;
@@ -60,7 +76,7 @@ import com.google.common.collect.Multimap;
 /**
  * An {@link ICDOConflictProcessor} implementation handling conflicts specific to the SNOMED CT terminology model.
  */
-public class SnomedCDOConflictProcessor extends AbstractCDOConflictProcessor implements ICDOConflictProcessor {
+public class SnomedCDOConflictProcessor extends AbstractCDOConflictProcessor {
 
 	private static final Set<EClass> COMPONENT_CLASSES = ImmutableSet.of(
 			SnomedPackage.Literals.CONCEPT, 
@@ -89,12 +105,8 @@ public class SnomedCDOConflictProcessor extends AbstractCDOConflictProcessor imp
 
 	private static final Logger LOGGER = LoggerFactory.getLogger(SnomedCDOConflictProcessor.class);
 	
-	private Map<String, CDOID> newComponentIdsInSource;
-	private Set<CDOID> detachedSourceIds;
 	private Map<String, CDOID> newComponentIdsInTarget;
 	private Set<CDOID> detachedTargetIds;
-
-	private boolean isRebase;
 
 	private Multimap<CDOID, Pair<EStructuralFeature, CDOID>> newSourceRevisionIdToFeatureIdMap;
 	private Multimap<CDOID, Pair<EStructuralFeature, CDOID>> newTargetRevisionIdToFeatureIdMap;
@@ -115,71 +127,30 @@ public class SnomedCDOConflictProcessor extends AbstractCDOConflictProcessor imp
 	 */
 	@Override
 	public Object addedInSource(final CDORevision sourceRevision, final Map<CDOID, Object> targetMap) {
-		if (isRebase) {
-			Conflict conflict = checkDuplicateComponentIds(sourceRevision, newComponentIdsInTarget);
-			
-			if (conflict != null) {
-				return conflict;
-			}
-			
-			conflict = checkDetachedReferences(sourceRevision.getID(), newSourceRevisionIdToFeatureIdMap.get(sourceRevision.getID()), detachedTargetIds);
-			
-			if (conflict != null) {
-				return conflict;
-			}
+		Conflict conflict = checkDuplicateComponentIds(sourceRevision, newComponentIdsInTarget);
+		
+		if (conflict != null) {
+			return conflict;
+		}
+		
+		conflict = checkDetachedReferences(sourceRevision.getID(), newSourceRevisionIdToFeatureIdMap.get(sourceRevision.getID()));
+		
+		if (conflict != null) {
+			return conflict;
 		}
 		
 		return super.addedInSource(sourceRevision, targetMap);
 	}
 	
 	@Override
-	public Object addedInTarget(final CDORevision targetRevision, final Map<CDOID, Object> sourceMap) {
-		
-		if (!isRebase) {
-			
-			Conflict conflict = checkDuplicateComponentIds(targetRevision, newComponentIdsInSource);
-			
-			if (conflict != null) {
-				return conflict;
-			}
-			
-			conflict = checkDetachedReferences(targetRevision.getID(), newTargetRevisionIdToFeatureIdMap.get(targetRevision.getID()), detachedSourceIds);
-			
-			if (conflict != null) {
-				return conflict;
-			}
-			
-		}
-		
-		return super.addedInTarget(targetRevision, sourceMap);
-	}
-	
-	@Override
 	public Object detachedInSource(CDOID id) {
+		Conflict conflict = checkDetachedReferences(id);
 		
-		if (isRebase) {
-			Conflict conflict = checkDetachedReferences(newTargetRevisionIdToFeatureIdMap, id);
-			
-			if (conflict != null) {
-				return conflict;
-			}
+		if (conflict != null) {
+			return conflict;
 		}
 		
 		return super.detachedInSource(id);
-	}
-	
-	@Override
-	public Object detachedInTarget(CDOID id) {
-		
-		if (!isRebase) {
-			Conflict conflict = checkDetachedReferences(newSourceRevisionIdToFeatureIdMap, id);
-			
-			if (conflict != null) {
-				return conflict;
-			}
-		}
-		
-		return super.detachedInTarget(id);
 	}
 	
 	@Override
@@ -201,31 +172,144 @@ public class SnomedCDOConflictProcessor extends AbstractCDOConflictProcessor imp
 	}
 	
 	@Override
-	public Collection<MergeConflict> handleCDOConflicts(final CDOView sourceView, final CDOView targetView, final Map<CDOID, Conflict> conflicts) {
-		if (!conflicts.isEmpty()) {
-			return FluentIterable.from(conflicts.values()).transform(new Function<Conflict, MergeConflict>() {
-				@Override public MergeConflict apply(Conflict input) {
-					return SnomedMergeConflictMapper.convert(input, sourceView, targetView);
-				}
-			}).toList();
-		}
-		return super.handleCDOConflicts(sourceView, targetView, conflicts);
+	protected MergeConflict convert(final ChangedInSourceAndTargetConflict conflict, final CDOView sourceView, final CDOView targetView) {
+		return MergeConflictImpl.builder()
+				.componentId(getComponentId(targetView, conflict.getTargetDelta().getID()))
+				.componentType(getType(targetView, conflict.getTargetDelta().getID()))
+				.conflictingAttributes(getConflictingAttributes(conflict.getTargetDelta(), targetView))
+				.type(ConflictType.CONFLICTING_CHANGE)
+				.build();
+	}
+
+	@Override
+	protected MergeConflict convert(final ChangedInSourceAndDetachedInTargetConflict conflict, final CDOView sourceView, final CDOView targetView) {
+		return MergeConflictImpl.builder()
+				.componentId(getComponentId(sourceView, conflict.getSourceDelta().getID()))
+				.componentType(getType(sourceView, conflict.getSourceDelta().getID()))
+				.conflictingAttributes(getConflictingAttributes(conflict.getSourceDelta(), sourceView))
+				.type(ConflictType.DELETED_WHILE_CHANGED)
+				.build();
+	}
+
+	@Override
+	protected MergeConflict convert(final ChangedInTargetAndDetachedInSourceConflict conflict, final CDOView sourceView, final CDOView targetView) {
+		return MergeConflictImpl.builder()
+				.componentId(getComponentId(targetView, conflict.getTargetDelta().getID()))
+				.componentType(getType(targetView, conflict.getTargetDelta().getID()))
+				.conflictingAttributes(getConflictingAttributes(conflict.getTargetDelta(), targetView))
+				.type(ConflictType.CHANGED_WHILE_DELETED)
+				.build();
+	}
+
+	@Override
+	protected MergeConflict convert(final AddedInSourceAndTargetConflict conflict, final CDOView sourceView, final CDOView targetView) {
+		return MergeConflictImpl.builder()
+				.componentId(getComponentId(targetView, conflict.getTargetId()))
+				.componentType(getType(targetView, conflict.getTargetId()))
+				.conflictingAttribute(ConflictingAttributeImpl.builder().property("id").build())
+				.type(ConflictType.CONFLICTING_CHANGE)
+				.build();
+	}
+
+	@Override
+	protected MergeConflict convert(final AddedInSourceAndDetachedInTargetConflict conflict, final CDOView sourceView, final CDOView targetView) {
+		return MergeConflictImpl.builder()
+				.componentId(getComponentId(sourceView, conflict.getSourceId()))
+				.componentType(getType(sourceView, conflict.getSourceId()))
+				.type(ConflictType.CAUSES_MISSING_REFERENCE)
+				.build();
 	}
 	
 	@Override
-	public void preProcess(final Map<CDOID, Object> sourceMap, final Map<CDOID, Object> targetMap, boolean isRebase) {
-		this.isRebase = isRebase;
-		
+	protected MergeConflict convert(final AddedInTargetAndDetachedInSourceConflict conflict, final CDOView sourceView, final CDOView targetView) {
+		return MergeConflictImpl.builder()
+				.componentId(getComponentId(targetView, conflict.getTargetId()))
+				.componentType(getType(targetView, conflict.getTargetId()))
+				.conflictingAttributes(
+						Strings.isNullOrEmpty(conflict.getFeatureName()) ? Collections.<ConflictingAttribute> emptyList() : 
+							Collections.<ConflictingAttribute> singletonList(ConflictingAttributeImpl.builder().property(conflict.getFeatureName()).build()))
+				.type(ConflictType.HAS_MISSING_REFERENCE)
+				.build();
+	}
+
+	private static String getComponentId(final CDOView view, final CDOID id) {
+		try {
+			final CDOObject object = view.getObject(id);
+			if (object != null) {
+				if (object instanceof Component) {
+					return ((Component) object).getId();
+				} else if (object instanceof SnomedRefSetMember) {
+					return ((SnomedRefSetMember) object).getUuid();
+				}
+			}
+		} catch (final ObjectNotFoundException e) {
+			// fall through
+		}
+		return id.toString();
+	}
+
+	private static String getType(final CDOView view, final CDOID id) {
+		try {
+			final CDOObject object = view.getObject(id);
+			return object.eClass().getName();
+		} catch (final ObjectNotFoundException e) {
+			// fall through
+		}
+		return null;
+	}
+	
+	private static List<ConflictingAttribute> getConflictingAttributes(final CDORevisionDelta cdoRevisionDelta, final CDOView view) {
+		return FluentIterable.from(cdoRevisionDelta.getFeatureDeltas()).transform(new Function<CDOFeatureDelta, ConflictingAttribute>() {
+			@Override 
+			public ConflictingAttribute apply(CDOFeatureDelta featureDelta) {
+				
+				String property = featureDelta.getFeature().getName();
+				String oldValue = null;
+				String newValue = null;
+				
+				if (featureDelta instanceof CDOSetFeatureDelta) {
+					CDOSetFeatureDelta setFeatureDelta = (CDOSetFeatureDelta) featureDelta;
+					newValue = convertValue(setFeatureDelta.getValue(), view);
+					oldValue = convertValue(setFeatureDelta.getOldValue(), view);
+				}
+				return ConflictingAttributeImpl.builder()
+							.property(property)
+							.oldValue(oldValue)
+							.value(newValue)
+							.build();
+			}
+
+		}).toList();
+	}
+
+	private static String convertValue(final Object value, final CDOView view) {
+		if (value instanceof CDOID) {
+			// try to resolve id
+			CDOID cdoId = (CDOID) value;
+			String componentId = getComponentId(view, cdoId);
+			if (!componentId.equals(cdoId.toString())) {
+				return componentId;
+			}
+		} else {
+			if (value instanceof Date) {
+				return Dates.formatByHostTimeZone(value, DateFormats.SHORT);
+			} else if (value != null) {
+				return value.toString();
+			}
+		}
+		return null;
+	}
+
+	
+	@Override
+	public void preProcess(final Map<CDOID, Object> sourceMap, final Map<CDOID, Object> targetMap) {
 		Collection<InternalCDORevision> newSourceComponentRevisions = extractNewComponentRevisions(sourceMap);
 		Collection<InternalCDORevision> newTargetComponentRevisions = extractNewComponentRevisions(targetMap);
 
 		newSourceRevisionIdToFeatureIdMap = extractNewRevisionIdToFeatureIdMap(newSourceComponentRevisions);
 		newTargetRevisionIdToFeatureIdMap = extractNewRevisionIdToFeatureIdMap(newTargetComponentRevisions);
 		
-		newComponentIdsInSource = extractNewComponentIds(newSourceComponentRevisions);
 		newComponentIdsInTarget = extractNewComponentIds(newTargetComponentRevisions);
-		
-		detachedSourceIds = getDetachedIdsInTarget(sourceMap);
 		detachedTargetIds = getDetachedIdsInTarget(targetMap);
 	}
 	
@@ -299,28 +383,20 @@ public class SnomedCDOConflictProcessor extends AbstractCDOConflictProcessor imp
 		return (String) revision.getValue(SnomedPackage.Literals.COMPONENT__ID);
 	}
 
-	private Conflict checkDetachedReferences(final CDOID revisionId, final Collection<Pair<EStructuralFeature, CDOID>> featureIds, final Set<CDOID> detachedIds) {
+	private Conflict checkDetachedReferences(final CDOID addedSourceId, final Collection<Pair<EStructuralFeature, CDOID>> featureIds) {
 		for (Pair<EStructuralFeature, CDOID> featureAndId : featureIds) {
-			if (detachedIds.contains(featureAndId.getB())) {
-				if (isRebase) {
-					return new AddedInSourceAndDetachedInTargetConflict(revisionId, featureAndId.getB(), featureAndId.getA().getName());
-				} else {
-					return new AddedInTargetAndDetachedInSourceConflict(featureAndId.getB(), revisionId, featureAndId.getA().getName());
-				}
+			if (detachedTargetIds.contains(featureAndId.getB())) {
+				return new AddedInSourceAndDetachedInTargetConflict(addedSourceId, featureAndId.getB(), featureAndId.getA().getName());
 			}
 		}
 		return null;
 	}
 	
-	private Conflict checkDetachedReferences(Multimap<CDOID, Pair<EStructuralFeature, CDOID>> newRevisionIdToFeatureIdMap, CDOID id) {
-		for (Entry<CDOID, Collection<Pair<EStructuralFeature, CDOID>>> entry : newRevisionIdToFeatureIdMap.asMap().entrySet()) {
+	private Conflict checkDetachedReferences(CDOID detachedSourceId) {
+		for (Entry<CDOID, Collection<Pair<EStructuralFeature, CDOID>>> entry : newTargetRevisionIdToFeatureIdMap.asMap().entrySet()) {
 			for (Pair<EStructuralFeature, CDOID> featureAndId : entry.getValue()) {
-				if (featureAndId.getB().equals(id)) {
-					if (isRebase) {
-						return new AddedInTargetAndDetachedInSourceConflict(id, entry.getKey(), featureAndId.getA().getName());
-					} else {
-						return new AddedInSourceAndDetachedInTargetConflict(entry.getKey(), id, featureAndId.getA().getName());
-					}
+				if (featureAndId.getB().equals(detachedSourceId)) {
+					return new AddedInTargetAndDetachedInSourceConflict(detachedSourceId, entry.getKey(), featureAndId.getA().getName());
 				}
 			}
 		}
